refactor(`MerkleTree`): Change index type back to usize

+++ twenty-first/src/util_types/mmr/mmr_membership_proof.rs
@@ -69,7 +69,7 @@ impl MmrMembershipProof {
             };
             mt_index /= 2;
         }
-        debug_assert_eq!(merkle_tree::ROOT_INDEX, mt_index);
+        debug_assert_eq!(merkle_tree::ROOT_INDEX as u64, mt_index);
 
         let peak_index = usize::try_from(peak_index).expect(U32_TO_USIZE_ERR);
